Bind player entity system to InputActions

+++ EngineCore/MouseMovedEvent.cpp
@@ -6,7 +6,7 @@
 #include "MouseMovedEvent.h"
 
 namespace core {
-	MouseMovedEvent::MouseMovedEvent(int xPos, int yPos) : Position(xPos, yPos) {
+	MouseMovedEvent::MouseMovedEvent(int xPos, int yPos) : Position((float)xPos, (float)yPos) {
 	}
 
 	utils::Name MouseMovedEvent::GetEventType() const {
